+++ src/test/polymer/iron-flex-layout-classes_test.ts
@@ -14,11 +14,10 @@
 
 import {assert} from 'chai';
 import * as path from 'path';
-import {Analyzer, FSUrlLoader} from 'polymer-analyzer';
+import {Analyzer, applyEdits, FSUrlLoader, makeParseLoader} from 'polymer-analyzer';
 
 import {Linter} from '../../linter';
 import {registry} from '../../registry';
-import {applyEdits, makeParseLoader} from '../../warning';
 import {WarningPrettyPrinter} from '../util';
 
 const fixtures_dir = path.join(__dirname, '..', '..', '..', 'test');
@@ -38,7 +37,7 @@ suite(ruleId, () => {
 
   test('works in the trivial case', async() => {
     const warnings = await linter.lint([]);
-    assert.deepEqual(warnings, []);
+    assert.deepEqual([...warnings], []);
   });
 
   test('warns for the proper cases and with the right messages', async() => {
@@ -72,6 +71,12 @@ suite(ruleId, () => {
           <div class="layout horizontal">
                      ~~~~~~~~~~~~~~~~~~~`,
       `
+      <div class$="layout horizontal [[flex]]">
+                  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~`,
+      `
+        <p class="flex" class$="flex [[flex]]">lorem</p>
+                               ~~~~~~~~~~~~~~~`,
+      `
 <body class="fullbleed">
             ~~~~~~~~~~~`,
     ]);
@@ -94,6 +99,10 @@ Import it in the template style include.`,
       'Style outside template. Run `move-style-into-template` rule.',
       `"iron-flex" style module is used but not imported.
 Import it in the template style include.`,
+      `"iron-flex" style module is used but not imported.
+Import it in the template style include.`,
+      `"iron-flex" style module is used but not imported.
+Import it in the template style include.`,
       `"iron-positioning" style module is used but not imported.
 Import it in the template style include.`,
     ]);
@@ -102,7 +111,7 @@ Import it in the template style include.`,
   test('applies automatic-safe fixes', async() => {
     const warnings = await linter.lint([`${ruleId}/before-fixes.html`]);
     const edits = warnings.filter((w) => w.fix).map((w) => w.fix!);
-    const loader = makeParseLoader(analyzer);
+    const loader = makeParseLoader(analyzer, warnings.analysis);
     const result = await applyEdits(edits, loader);
     assert.deepEqual(
         result.editedFiles.get(`${ruleId}/before-fixes.html`),