+++ src/test/polymer/iron-flex-layout-import_test.ts
@@ -14,11 +14,10 @@
 
 import {assert} from 'chai';
 import * as path from 'path';
-import {Analyzer, FSUrlLoader} from 'polymer-analyzer';
+import {Analyzer, applyEdits, FSUrlLoader, makeParseLoader} from 'polymer-analyzer';
 
 import {Linter} from '../../linter';
 import {registry} from '../../registry';
-import {applyEdits, makeParseLoader} from '../../warning';
 import {WarningPrettyPrinter} from '../util';
 
 const fixtures_dir = path.join(__dirname, '..', '..', '..', 'test');
@@ -38,7 +37,7 @@ suite(ruleId, () => {
 
   test('works in the trivial case', async() => {
     const warnings = await linter.lint([]);
-    assert.deepEqual(warnings, []);
+    assert.deepEqual([...warnings], []);
   });
 
   test('warns when deprecated files are used with the right messages', async() => {
@@ -69,7 +68,7 @@ Run the lint rule \`iron-flex-layout-classes\` with \`--fix\` to include the req
         const warnings =
             await linter.lint([`${ruleId}/deprecated-files-before-fixes.html`]);
         const edits = warnings.filter((w) => w.fix).map((w) => w.fix!);
-        const loader = makeParseLoader(analyzer);
+        const loader = makeParseLoader(analyzer, warnings.analysis);
         const result = await applyEdits(edits, loader);
         assert.deepEqual(
             result.editedFiles.get(
@@ -99,7 +98,7 @@ Import iron-flex-layout/iron-flex-layout-classes.html`,
     const warnings =
         await linter.lint([`${ruleId}/forgot-import-before-fixes.html`]);
     const edits = warnings.filter((w) => w.fix).map((w) => w.fix!);
-    const loader = makeParseLoader(analyzer);
+    const loader = makeParseLoader(analyzer, warnings.analysis);
     const result = await applyEdits(edits, loader);
     assert.deepEqual(
         result.editedFiles.get(`${ruleId}/forgot-import-before-fixes.html`),
@@ -110,7 +109,7 @@ Import iron-flex-layout/iron-flex-layout-classes.html`,
     const warnings = await linter.lint(
         [`${ruleId}/forgot-import-no-imports-before-fixes.html`]);
     const edits = warnings.filter((w) => w.fix).map((w) => w.fix!);
-    const loader = makeParseLoader(analyzer);
+    const loader = makeParseLoader(analyzer, warnings.analysis);
     const result = await applyEdits(edits, loader);
     assert.deepEqual(
         result.editedFiles.get(
@@ -137,7 +136,7 @@ Import iron-flex-layout/iron-flex-layout-classes.html`,
     const warnings =
         await linter.lint([`${ruleId}/unnecessary-import-before-fixes.html`]);
     const edits = warnings.filter((w) => w.fix).map((w) => w.fix!);
-    const loader = makeParseLoader(analyzer);
+    const loader = makeParseLoader(analyzer, warnings.analysis);
     const result = await applyEdits(edits, loader);
     assert.deepEqual(
         result.editedFiles.get(